+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+### 3.9.1(33)
+
+Localization Service
+
+- `LocalizationService` no longer causes issues with default localization outside of non-English regions
+
 ### [3.9.0(32)](https://github.com/VakhoKontridze/VCore/releases/tag/3.9.0) — *2022 06 27*
 
 Extensions